LayoutTests/webarchive/archive-empty-frame-source.html

<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN"
        "http://www.w3.org/TR/html4/loose.dtd">
<html>
<head>
<title>Empty iFrame causes archive exception</title>
<script>
    if (window.testRunner)
        testRunner.dumpSourceAsWebArchive();
</script>
</head>
<body>

<p>
Here lies an `iframe` with no content/src:
<p>

<iframe id="emptyFrame" style="width: 100%; height: 100px; border: solid 2px red;">
</iframe>

</body>
</html>
